feat(plasma-new-hope): resolve circular dependency in components
