+++ packages/plasma-new-hope/src/components/Combobox/ComboboxNew/Combobox.context.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import type { ItemContext } from './Combobox.types';
+
+export const Context = createContext<ItemContext>({} as ItemContext);
