+++ packages/plasma-new-hope/src/components/Dropdown/Dropdown.context.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import type { ItemContext } from './Dropdown.types';
+
+export const Context = createContext<ItemContext>({} as ItemContext);
