+++ packages/plasma-new-hope/src/components/Select/Select.context.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import type { ItemContext } from './Select.types';
+
+export const Context = createContext<ItemContext>({} as ItemContext);
